// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/content-report-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ContentReportHttpResponse {
    adminNote?: string | null;
    createdAtUtc?: string;
    id?: string;
    reason?: string;
    reporterId?: string;
    reviewedAtUtc?: string | null;
    status?: string;
    targetId?: string;
    targetType?: string;
}